Move tasks/accounts.ts to hardhat-scripts/utils
