+++ hardhat.config.ts
@@ -14,7 +14,7 @@ import type {
 import { resolve } from "path";
 import fs from "fs";
 
-import "./tasks/accounts";
+import "./hardhat-scripts/utils/accounts";
 import { getJsonRpcUrl } from "./hardhat-scripts/constants/networks";
 import {
   ChainId,
